Added property_scraper base-class skeleton
